Fix bug in karrerNewman.R
Previous algorithm contained a mistake, it did not correspond to the algorithm described in the paper by Karrer and Newman. It would sample uniformly from the set of vertices with in-stubs at each step, instead of sampling uniformly from the set of in-stubs. This is fixed now. The current algorithm does correspond to the configuration model described in the paper.